requirements.txt
ultralytics
Pillow
opencv-python
mss
numpy
